Search

+++ CMPUT301W15T04Team04Project/src/ca/ualberta/cs/cmput301w15t04team04project/models/Currency.java
@@ -51,7 +51,7 @@ public class Currency {
 		this.type = type;
 	}
 
-	public double getAmount() {
+	public int getAmount() {
 		return amount;
 	}
